// component/profile/type.ts
export type MyDetailType = {
  displayName: string;
  email: string;
  password: string;
};

export type PasswordChangeType = {
  oldPassword: string;
  newPassword: string;
};

export type GeneralDetailType = {
  monthlySalary: string;
  monthlySavings: string;
  currentCapital: string;
  country: string;
  age: string;
  gender: string;
  prefferedCurrency: string;
};